adding a Jekyll build action
### effect of this change:  

When a commit is pushed to the `gh-pages` branch Google will run a Jekyll build and one can see the results in the _Actions_ tab.  This will allow you to see any errors that might occur.  

Previously the only way to see Jekyll build errors was to have Jekyll installed locally and do the build on your own computer watching the command line console for errors.

Hopefully this will make troubleshooting Jekyll errors easier.